\name{hgu133abarcodevecs}
\docType{data}
\alias{hgu133abarcodevecs}
\title{hgu133a data for barcode}
\description{Data used by the barcode package for microarrays of type hgu133a.}
\usage{
  data(bcparams-GPL96)
}
\format{
  A matrix of parameters estimates used in the barcode function for
  Affymetrix hgu133a microarray data.
}
\keyword{datasets}



